Loads of changes:
Renamed MinuitFunction to Minuit2Function

Started implementing the "fail on any error" approach. This mean InvalidObject is useless, and has been removed.

Added features to the Minimiser and FitFunction xml tags, allowing the specification of requested contour plots and event weights respectively.
This required the introduction of two new objects FitFunctionConfiguration and MinimiserConfiguration to hold this (and future) information.

Changed Greig's contour plotting code to perform multiple (or no) contour plots, over requested variables. This required changes to
MinuitWrapper, Minuit2Wrapper, and ResultFormatter. Also introduced the FunctionContour object to make the data easier to use.

NOTE: Minuit2 is much slower with the contour plots than Minuit, because it insists on re-calculating the minimum on change of error. May be able
to improve this, or at least optimise to it.

NOTE: This is all reasonably well tested, but only with the JPsiPhi PDF. RaPDFToy_noBkg.xml implements the new code.